moved command line processing to cli.py
