+++ src/xacro/__init__.py
@@ -38,14 +38,13 @@ import re
 import sys
 import ast
 import math
-import textwrap
 
 from roslaunch import substitution_args
-from rosgraph.names import load_mappings, REMAP
-from optparse import OptionParser, IndentedHelpFormatter
 from copy import deepcopy
-from .color import warning, error, message, colorize
+from .color import warning, error, message
 from .xmlutils import *
+from .cli import process_args
+
 
 try:
     _basestr = basestring
@@ -793,80 +792,6 @@ def eval_all(node, macros, symbols):
 
         node = next
 
-class ColoredOptionParser(OptionParser):
-    def error(self, message):
-        msg = colorize(message, 'red')
-        OptionParser.error(self, msg)
-
-
-_original_wrap = textwrap.wrap
-def wrap_with_newlines(text, width, **kwargs):
-    result = []
-    for paragraph in text.split('\n'):
-        result.extend(_original_wrap(paragraph, width, **kwargs))
-    return result
-
-class IndentedHelpFormatterWithNL(IndentedHelpFormatter):
-    def __init__(self, *args, **kwargs):
-        IndentedHelpFormatter.__init__(self, *args, **kwargs)
-
-    def format_option(self, text):
-        textwrap.wrap, old = wrap_with_newlines, textwrap.wrap
-        result = IndentedHelpFormatter.format_option(self, text)
-        textwrap.wrap = old
-        return result
-
-
-def process_cli_args(argv, require_input=True):
-    parser = ColoredOptionParser(usage="usage: %prog [options] <input>",
-                                 formatter=IndentedHelpFormatterWithNL())
-    parser.add_option("-o", dest="output", metavar="FILE",
-                      help="write output to FILE instead of stdout")
-    parser.add_option("--oldorder", action="store_false", dest="in_order",
-                      help="use traditional processing order [deprecated default]")
-    parser.add_option("--inorder", action="store_true", dest="in_order",
-                      help="use processing in read order")
-    parser.add_option("--deps", action="store_true", dest="just_deps",
-                      help="print file dependencies")
-    parser.add_option("--includes", action="store_true", dest="just_includes",
-                      help="only process includes")
-    parser.add_option("--xacro-ns", action="store_false", default=True, dest="xacro_ns",
-                      help="require xacro namespace prefix for xacro tags")
-
-    # verbosity options
-    parser.add_option("-q", action="store_const", dest="verbosity", const=0,
-                      help="quiet operation suppressing warnings")
-    parser.add_option("-v", action="count", dest="verbosity",
-                      help="increase verbosity")
-    parser.add_option("--verbosity", metavar='level', dest="verbosity", type='int',
-                      help=textwrap.dedent("""\
-                      set verbosity level
-                      0: quiet, suppressing warnings
-                      1: default, showing warnings and error locations
-                      2: show stack trace
-                      3: log property definitions and usage on top level
-                      4: log property definitions and usage on all levels"""))
-
-    # process substitution args
-    mappings = load_mappings(argv)
-
-    parser.set_defaults(in_order=False, just_deps=False, just_includes=False, 
-                        verbosity=verbosity)
-    filtered_args = [a for a in argv if REMAP not in a]  # filter-out REMAP args
-    (options, pos_args) = parser.parse_args(filtered_args)
-
-    if options.in_order and options.just_includes:
-        parser.error("options --inorder and --includes are mutually exclusive")
-
-    if len(pos_args) != 1:
-        if require_input:
-            parser.error("expected exactly one input file as argument")
-        else:
-            pos_args = [None]
-
-    options.mappings = mappings
-    return options, pos_args[0]
-
 
 def parse(inp, filename=None):
     """
@@ -967,7 +892,7 @@ def print_location(filestack, err=None, file=sys.stderr):
 
 
 def main():
-    opts, input_file = process_cli_args(sys.argv[1:])
+    opts, input_file = process_args(sys.argv[1:])
     global verbosity
     verbosity = opts.verbosity
